requirements.txt
# Web Server & API
fastapi[standard]
uvicorn
pydantic

# Blockchain
solana==0.35.0
solders==0.21.0
anchorpy==0.20.1

# AI & LLM (Ollama handles the A770 natively)
google-generativeai
ollama

# Utility
python-dotenv
requests
aiohttp
pandas
matplotlib

# Testing
pytest
httpx
